fix : globalPos가 바뀔 때 움직이는 오브젝트들의 움직임 수정
움직일 때 globalPos를 int형으로 변환하는 부분 삭제

+++ SpinCoin.cpp
@@ -7,7 +7,7 @@ void SpinCoin::Move()
 {
     if (PLAYER->GetPos().x + PLAYER->GetCurrSpeed() > WIN_SIZE_X / 2)
     {
-        pos.x -= (int)PLAYER->GetCurrSpeed();
+        pos.x -= PLAYER->GetCurrSpeed();
     }
 }
 void SpinCoin::ChangeFrame()
